fix: resolve all clippy warnings across non-TUI crates
- Auto-fix redundant closures, useless vec!, borrowed expressions
- Replace assert_eq!(x, false/true) with assert!(!x)/assert!(x)
- Use struct init syntax instead of field reassignment after Default
- Allow await_holding_lock in test modules using ENV_MUTEX for env serialization
- Remove duplicated #[allow(dead_code)] attribute on ToastLevel

+++ crates/at-cli/src/commands/smoke.rs
@@ -84,7 +84,7 @@ pub async fn run(
 ) -> anyhow::Result<()> {
     let parsed = Url::parse(ui_url).context("invalid --ui-url")?;
     let host = parsed.host_str().unwrap_or("127.0.0.1").to_string();
-    let port = parsed.port_or_known_default().unwrap_or(3001) as u16;
+    let port = parsed.port_or_known_default().unwrap_or(3001);
 
     let mut static_server = None;
     if serve_dist {
